reorganize waveforms and fetaures folder
