+++ src/spikeinterface/sortingcomponents/waveforms/peak_svd.py
