+++ src/spikeinterface/sortingcomponents/waveforms/tests/conftest.py
