+++ src/spikeinterface/sortingcomponents/waveforms/tests/test_hanning_filter.py
